requirements.txt
django
djangorestframework
celery
# redis (necesito instalar Linux para este)
psycopg2